test: change run mongo
