+++ module/data/src/main/kotlin/io/github/siyual_park/data/test/ResourceTestHelper.kt
@@ -1,0 +1,6 @@
+package io.github.siyual_park.data.test
+
+interface ResourceTestHelper {
+    fun setUp()
+    fun tearDown()
+}
